# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级后台管理系统接口文档，覆盖认证、用户、角色、菜单、部门、字典和审计等业务模块。",
        "title": "Go Server 后台管理 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/system/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回完整菜单树，常用于菜单管理页面展示。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "查询菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_menu.MenuTreeNode"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建目录、菜单或按钮类型的权限节点。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "菜单创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_menu.CreateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_menu.MenuResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/menus/current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据当前登录用户的权限返回可见菜单树。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "查询当前用户菜单",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_menu.MenuTreeNode"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/menus/sort": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批量更新多个菜单节点的排序值。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "更新菜单排序",
                "parameters": [
                    {
                        "description": "菜单排序更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_menu.UpdateMenuSortsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/menus/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据菜单 ID 查询菜单详情。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_menu.MenuResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改菜单节点的名称、路径、组件、权限标识和显示状态等信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_menu.UpdateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_menu.MenuResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据菜单 ID 删除菜单节点。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
...
        "/system/roles/{id}/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询角色已勾选的菜单 ID 及完整菜单树结构。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "获取角色菜单权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_role.RoleMenusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为角色分配菜单权限。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "更新角色菜单权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单 ID 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_role.AssignMenusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-server_internal_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_modules_menu.CreateMenuRequest": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "component": {
                    "type": "string",
                    "example": "/system/user/index"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "用户管理"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "path": {
                    "type": "string",
                    "example": "/system/user"
                },
                "permission": {
                    "type": "string",
                    "example": "system:user:list"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "directory",
                        "menu",
                        "button"
                    ],
                    "example": "menu"
                },
                "visible": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                }
            }
        },
        "internal_modules_menu.MenuResponse": {
            "type": "object",
            "properties": {
                "component": {
                    "type": "string",
                    "example": "/system/user/index"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-11T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "用户管理"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "path": {
                    "type": "string",
                    "example": "/system/user"
                },
                "permission": {
                    "type": "string",
                    "example": "system:user:list"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "example": "menu"
                },
                "visible": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_modules_menu.MenuSortItem": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "sort": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "internal_modules_menu.MenuTreeNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_menu.MenuTreeNode"
                    }
                },
                "component": {
                    "type": "string",
                    "example": "/system/index"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "系统管理"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "path": {
                    "type": "string",
                    "example": "/system"
                },
                "permission": {
                    "type": "string",
                    "example": "system:user:list"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "example": "directory"
                },
                "visible": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_modules_menu.UpdateMenuRequest": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "component": {
                    "type": "string",
                    "example": "/system/user/index"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "用户管理"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "path": {
                    "type": "string",
                    "example": "/system/user"
                },
                "permission": {
                    "type": "string",
                    "example": "system:user:list"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "directory",
                        "menu",
                        "button"
                    ],
                    "example": "menu"
                },
                "visible": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                }
            }
        },
        "internal_modules_menu.UpdateMenuSortsRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/internal_modules_menu.MenuSortItem"
                    }
                }
            }
        },
...
        "internal_modules_role.AssignMenusRequest": {
            "type": "object",
            "properties": {
                "menu_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                }
            }
        },
...
        "internal_modules_role.RoleMenusResponse": {
            "type": "object",
            "properties": {
                "checked_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_role.MenuTreeNode"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请填写 Bearer Token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}